pyproject.toml
[project]
name = "mcp_lg_tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=3.4.3",
    "langchain==0.3.25",
    "langchain-community==0.3.24",
    "langchain-core==0.3.60",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-openai==0.3.17",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.4.5",
    "langgraph-checkpoint==2.0.26",
    "mcp[cli]>=1.6.0",
    "openai>=1.57.0",
    "python-dotenv>=1.0.1",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
]
